complete lab-authorization-basic

+++ Branche technique/Labs/lab-authorization-basic/app/Http/Controllers/TaskController.php
@@ -9,12 +9,11 @@ use
     use App\Repositories\TasksRepository;
 
 
-    
+
 class TaskController extends Controller
 {
 
     protected $tasksRepository;
-    // protected $projectRepository;
 
     public function __construct(TasksRepository $tasksRepository){
         $this->tasksRepository = $tasksRepository;
@@ -34,11 +33,13 @@ class TaskController extends Controller
 
     public function create(Project $project)
     {
+        $this->authorize('create', Task::class);
         return view('task.create', compact('project'));
     }
 
     public function store(Request $request, Project $project)
     {
+        $this->authorize('store', Task::class);
         $data = $request->validate([
             'name' => 'required',
             'description' => 'required',
@@ -53,6 +54,7 @@ class TaskController extends Controller
 
     public function edit(Project $project, Task $task)
     {
+        $this->authorize('edit', $task);
         $task = $this->tasksRepository->edit($task);
         return view('task.edit', ['task' => $task, 'project' => $project]);
     }
@@ -61,6 +63,7 @@ class TaskController extends Controller
 
     public function update(Request $request, Project $project,Task $task)
     {
+        $this->authorize('update', $task);
         $data = $request->validate([
             'name' => 'required',
             'description' => 'required',
@@ -75,6 +78,7 @@ class TaskController extends Controller
 
     public function destroy(Project $project,Task $task)
     {
+        $this->authorize('destroy', $task);
         $this->tasksRepository->destroy($task);
 
         return redirect()->route('task.index', ['project' => $project])->with('success', 'tache supprimé avec succès');
@@ -86,26 +90,18 @@ class TaskController extends Controller
         $search = $request->input('search');
         $project = Project::findOrFail($project);
 
-        // Check if the search value is empty
-        // $tasksQuery = ['project_id' => $project->id];
-        // $tasksQuery = ['name' => $search];
-
         if (empty($search)) {
             $tasks = $this->tasksRepository->index(['project_id' => $project->id]);
-            // $tasks = Task::where('project_id',$project->id)->paginate(3);
         } else {
             $tasks =  $this->tasksRepository->index(['project_id' => $project->id, 'search' => $search]);
-            // $tasks = Task::where('project_id', $project->id)->where('name', 'like', '%' . $search . '%')->paginate(3);
         }
 
-        // Controller code
         if ($request->ajax()) {
             return response()->json([
                 'table' => view('task.table', compact('tasks', 'project'))->render(),
                 'pagination' => $tasks->links()->toHtml(),
             ]);
         }
-
     }
